WIP: move gpm/trmm reader to `io.hdf`, move gr-reader to notebook, change imports
